Changed guest id to long, changed GuestRepository to GuestFileRepository, extracted interface GuestRepository, added new class GuestDatabaseRepository and implemented CRUD.

+++ src/main/java/org/example/domain/guest/GuestService.java
@@ -40,11 +40,11 @@ public class GuestService {
         repository.readAll();
     }
 
-    public void removeGuest(int id) {
+    public void removeGuest(long id) {
         repository.remove(id);
     }
 
-    public void editGuest(int id, String firstName, String lastName, int age, boolean isMale) {
+    public void editGuest(long id, String firstName, String lastName, int age, boolean isMale) {
         Gender gender = Gender.FEMALE;
         if (isMale) {
             gender = Gender.MALE;
