+++ src/main/java/org/example/domain/reservation/Reservation.java
@@ -39,6 +39,6 @@ public class Reservation {
     public ReservationDTO getAsDTO() {
         return new ReservationDTO(this.id, this.from,
                 this.to, (int)this.room.getId(), this.room.getNumber(),
-                this.guest.getId(), this.guest.getFirstName() + " " + this.guest.getLastName());
+                (int)this.guest.getId(), this.guest.getFirstName() + " " + this.guest.getLastName());
     }
 }
